# api/openapi.yaml
openapi: 3.1.0
info:
  title: Lute public API
  version: "1"
  description: |
    The API that API keys reach: list jobs and start them, create and follow runs, and manage
    workers and their registration tokens. The `lute` command line is built on it. The panel's own
    API under `/api/v1` is internal and not covered here.
  license:
    name: MIT
    url: https://github.com/notna-k/Lute/blob/master/LICENSE
servers:
  - url: https://lute.example.com/api/public/v1
    description: Your Lute core
security:
  - apiKey: []
...
paths:
...
  /jobs:
    get:
      tags: [jobs]
      operationId: listJobs
      summary: List jobs
      description: Every job definition, each with the newest run the key can see.
      responses:
        "200":
          description: The jobs.
          content:
            application/json:
              schema:
                type: object
                required: [jobs]
                properties:
                  jobs:
                    type: array
                    items: { $ref: "#/components/schemas/JobSummary" }
              example:
                jobs:
                  - slug: web-release
                    name: Web release
                    description: Build and ship the web app
                    queue: deploy
                    runtime: node:25-alpine
                    last_run:
                      id: 66fb1c2d3e4f5a6b7c8d9e0f
                      job: web-release
                      queue: deploy
                      type: container
                      status: done
                      attempts: 1
                      max_retries: 3
                      timeout_sec: 300
                      enqueued_at: "2026-09-30T09:14:02Z"
        "401": { $ref: "#/components/responses/Unauthorized" }
  /jobs/{slug}:
    parameters:
      - $ref: "#/components/parameters/JobSlug"
    get:
      tags: [jobs]
      operationId: getJob
      summary: Get a job
      description: |
        The job and its parameter schema: types, defaults and options, which is what a client needs
        to know what to pass. A `secret` parameter is listed, but nothing about its value is.
      responses:
        "200":
          description: The job.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/JobDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: "An API key from Settings → API keys, sent as `Authorization: Bearer <key>`."
  parameters:
...
    JobSlug:
      name: slug
      in: path
      required: true
      schema: { type: string }
      description: The job's `slug`.
      example: web-release
...
  responses:
...
    Unauthorized:
      description: Missing, unknown or revoked API key.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: unauthorized, message: invalid API key }
    NotFound:
      description: Not found, or not yours.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: not_found, message: run not found }
...
  schemas:
...
    JobSummary:
      type: object
      required: [slug, name, queue, runtime]
      properties:
        slug: { type: string }
        name: { type: string }
        description: { type: string }
        queue: { type: string, description: Workers serving this queue run it. }
        runtime: { type: string, description: The image its command runs in. }
        last_run: { $ref: "#/components/schemas/Run" }
    JobDetail:
      allOf:
        - $ref: "#/components/schemas/JobSummary"
        - type: object
          required: [command, parameters]
          properties:
            command: { type: string }
            source_repo: { type: string, description: Git URL checked out into the workspace. }
            parameters:
              type: array
              items: { $ref: "#/components/schemas/Parameter" }
      example:
        slug: web-release
        name: Web release
        description: Build and ship the web app
        queue: deploy
        runtime: node:25-alpine
        command: ./scripts/ship.sh
        parameters:
          - name: environment
            type: select
            label: Environment
            env_var: ENVIRONMENT
            required: true
            options:
              - { value: staging, label: Staging }
              - { value: prod, label: Production }
          - { name: dry_run, type: bool, label: Dry run, env_var: DRY_RUN, required: false, default: false }
    Parameter:
      type: object
      required: [name, type, env_var, required]
      properties:
        name: { type: string, description: The key to use in `params`. }
        type:
          type: string
          enum: [string, number, bool, select, multiselect, date, datetime, secret]
        label: { type: string }
        env_var: { type: string, description: The environment variable the value reaches the job as. }
        description: { type: string }
        required: { type: boolean }
        default: { description: Used when the value is left out. Never set for a secret. }
        options:
          type: array
          description: The allowed values of a `select` or `multiselect`.
          items:
            type: object
            required: [value]
            properties:
              value: { type: string }
              label: { type: string }
              hint: { type: string }
...
    Run:
      type: object
      required: [id, queue, type, status, attempts, max_retries, timeout_sec, enqueued_at]
      properties:
        id: { type: string, description: "24 hex characters. The first 8 are the short form, `#66fb1c2d`." }
        job: { type: string, description: "The job's slug, for a run started from a job." }
        params:
          type: object
          additionalProperties: { type: string }
          description: The values a job run got, keyed by env var. Secrets are never included.
        queue: { type: string }
        type: { type: string }
        status: { $ref: "#/components/schemas/RunStatus" }
        priority: { type: number }
        attempts: { type: integer }
        max_retries: { type: integer }
        timeout_sec: { type: integer }
        error: { type: string, description: "Last failure, if any." }
        worker_id: { type: string, description: Set once a worker leased the run. }
        idempotency_key: { type: string }
        webhook_url: { type: string }
        webhook_events: { type: array, items: { type: string } }
        enqueued_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time }
        finished_at: { type: string, format: date-time }
        elapsed_ms: { type: integer }
      example:
        id: 66fb1c2d3e4f5a6b7c8d9e0f
        queue: deploy
        type: container
        status: running
        attempts: 1
        max_retries: 2
        timeout_sec: 900
        worker_id: 66f9e0a1c4b2d3e4f5a6b7c8
        idempotency_key: release-2026-09-30
        webhook_url: https://ci.example.com/hooks/lute
        webhook_events: [run.completed, run.failed]
        enqueued_at: "2026-09-30T09:14:02Z"
        started_at: "2026-09-30T09:14:03Z"
...
    RunStatus:
      type: string
      enum: [pending, running, done, failed, dead, unknown]
      description: |
        `pending` until a worker takes it, then `running`. `done` and `failed` are final; `dead`
        means it ran out of retries or was cancelled.
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [bad_request, validation_failed, unauthorized, forbidden, not_found, conflict, unavailable, internal]
            message: { type: string }
            fields:
              type: object
              additionalProperties: { type: string }
              description: Per-input messages, for `validation_failed`.